% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_eml.R
\name{render_eml}
\alias{render_eml}
\title{Render EML}
\usage{
render_eml(file, open = TRUE, outfile = "test.html")
}
\arguments{
\item{file}{A valid Ecological Metadata Language file to be rendered to html.}

\item{open}{Whether to open the file in a browser. Defaults to TRUE.}

\item{outfile}{Name of output file.}
}
\value{
HTML file containing dataset information
}
\description{
Render EML metadata into a webpage
}
\details{
Pass in an xml file of EML metadata and generate a nice webpage describing
the dataset.
}
\examples{
render_eml(system.file("extdata", "Rodents_snakes_and_raptors.xml", package = "emldown"))
}
\author{
Kara Woo
}
