% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_eml.R
\name{write_custom_css}
\alias{write_custom_css}
\title{Write a custom css file}
\usage{
write_custom_css(publish_mode)
}
\arguments{
\item{publish_mode}{TRUE. If TRUE the website is pretty without warnings for 
weird stuff.}
}
\value{
Prints a custom.css file into the root directory.
}
\description{
The custom.css files make some small changes to the Bootstrap template.
Currently this is written into the root directory of your project.
}
